+++ crates/egui/src/widgets/drag_value.rs
@@ -461,23 +461,31 @@ impl<'a> Widget for DragValue<'a> {
         // some clones below are redundant if AccessKit is disabled
         #[allow(clippy::redundant_clone)]
         let mut response = if is_kb_editing {
-            let button_width = ui.spacing().interact_size.x;
             let mut value_text = ui
                 .memory_mut(|mem| mem.drag_value.edit_string.take())
                 .unwrap_or_else(|| value_text.clone());
             let response = ui.add(
                 TextEdit::singleline(&mut value_text)
+                    .clip_text(false)
+                    .horizontal_align(ui.layout().horizontal_align())
+                    .vertical_align(ui.layout().vertical_align())
+                    .margin(ui.spacing().button_padding)
+                    .min_size(ui.spacing().interact_size)
                     .id(id)
-                    .desired_width(button_width)
+                    .desired_width(ui.spacing().interact_size.x)
                     .font(text_style),
             );
-            let parsed_value = match custom_parser {
-                Some(parser) => parser(&value_text),
-                None => value_text.parse().ok(),
-            };
-            if let Some(parsed_value) = parsed_value {
-                let parsed_value = clamp_to_range(parsed_value, clamp_range.clone());
-                set(&mut get_set_value, parsed_value);
+            // Only update the value when the user presses enter, or clicks elsewhere. NOT every frame.
+            // See https://github.com/emilk/egui/issues/2687
+            if response.lost_focus() {
+                let parsed_value = match custom_parser {
+                    Some(parser) => parser(&value_text),
+                    None => value_text.parse().ok(),
+                };
+                if let Some(parsed_value) = parsed_value {
+                    let parsed_value = clamp_to_range(parsed_value, clamp_range.clone());
+                    set(&mut get_set_value, parsed_value);
+                }
             }
             ui.memory_mut(|mem| mem.drag_value.edit_string = Some(value_text));
             response
@@ -557,28 +565,28 @@ impl<'a> Widget for DragValue<'a> {
         response.widget_info(|| WidgetInfo::drag_value(value));
 
         #[cfg(feature = "accesskit")]
-        ui.ctx().accesskit_node(response.id, |node| {
+        ui.ctx().accesskit_node_builder(response.id, |builder| {
             use accesskit::Action;
             // If either end of the range is unbounded, it's better
             // to leave the corresponding AccessKit field set to None,
             // to allow for platform-specific default behavior.
             if clamp_range.start().is_finite() {
-                node.min_numeric_value = Some(*clamp_range.start());
+                builder.set_min_numeric_value(*clamp_range.start());
             }
             if clamp_range.end().is_finite() {
-                node.max_numeric_value = Some(*clamp_range.end());
+                builder.set_max_numeric_value(*clamp_range.end());
             }
-            node.numeric_value_step = Some(speed);
-            node.actions |= Action::SetValue;
+            builder.set_numeric_value_step(speed);
+            builder.add_action(Action::SetValue);
             if value < *clamp_range.end() {
-                node.actions |= Action::Increment;
+                builder.add_action(Action::Increment);
             }
             if value > *clamp_range.start() {
-                node.actions |= Action::Decrement;
+                builder.add_action(Action::Decrement);
             }
             // The name field is set to the current value by the button,
             // but we don't want it set that way on this widget type.
-            node.name = None;
+            builder.clear_name();
             // Always expose the value as a string. This makes the widget
             // more stable to accessibility users as it switches
             // between edit and button modes. This is particularly important
@@ -599,7 +607,7 @@ impl<'a> Widget for DragValue<'a> {
             // when in edit mode.
             if !is_kb_editing {
                 let value_text = format!("{}{}{}", prefix, value_text, suffix);
-                node.value = Some(value_text.into());
+                builder.set_value(value_text);
             }
         });