+++ examples/keyboard_events/src/main.rs
@@ -2,6 +2,7 @@
 
 use eframe::egui;
 use egui::*;
+
 fn main() -> Result<(), eframe::Error> {
     // Log to stdout (if you run with `RUST_LOG=debug`).
     tracing_subscriber::fmt::init();